+++ src/category-group/category-group.controller.ts
@@ -10,20 +10,19 @@ import {
 	Res,
 	Session
 } from '@nestjs/common'
-
-import { CategoryService } from './category.service'
-import { SessionData } from 'express-session'
 import { Response } from 'express'
-import { CreateCategoryDto } from './dto/create-category.dto'
-import { UpdateCategoryDto } from './dto/update-category.dto'
-import { ReorderCategoryDto } from './dto/reorder-category.dto'
+import { SessionData } from 'express-session'
+import { CategoryGroupService } from './category-group.service'
+import { CreateCategoryGroupDto } from './dto/create-category-group.dto'
+import { UpdateCategoryGroupDto } from './dto/update-category-group.dto'
+import { ReorderCategoryGroupDto } from './dto/reorder-category-group.dto'
 
-@Controller('category')
-export class CategoryController {
-	constructor(private categoryService: CategoryService) {}
+@Controller('group')
+export class CategoryGroupController {
+	constructor(private categoryGroupService: CategoryGroupService) {}
 
 	@Get('all')
-	async getAll(@Session() session: SessionData, @Res() res: Response) {
+	async all(@Session() session: SessionData, @Res() res: Response) {
 		if (session.auth === undefined || !session.auth.authorized) {
 			res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' })
 			return
@@ -31,14 +30,14 @@ export class CategoryController {
 
 		res
 			.status(HttpStatus.OK)
-			.json(await this.categoryService.getAll(session.auth.userId))
+			.json(await this.categoryGroupService.getAll(session.auth.userId))
 	}
 
 	@Post('add')
-	async addNew(
+	async add(
 		@Session() session: SessionData,
 		@Res() res: Response,
-		@Body() body: CreateCategoryDto
+		@Body() body: CreateCategoryGroupDto
 	) {
 		if (session.auth === undefined || !session.auth.authorized) {
 			res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' })
@@ -47,28 +46,12 @@ export class CategoryController {
 
 		res
 			.status(HttpStatus.OK)
-			.json(await this.categoryService.add(body, session.auth.userId))
-	}
-
-	@Delete(':id')
-	async delete(
-		@Param('id') id: string,
-		@Session() session: SessionData,
-		@Res() res: Response
-	) {
-		if (session.auth === undefined || !session.auth.authorized) {
-			res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' })
-			return
-		}
-
-		res
-			.status(HttpStatus.OK)
-			.json(await this.categoryService.delete(id, session.auth.userId))
+			.json(await this.categoryGroupService.add(body, session.auth.userId))
 	}
 
 	@Put('reorder')
 	async reorder(
-		@Body() body: ReorderCategoryDto,
+		@Body() body: ReorderCategoryGroupDto,
 		@Session() session: SessionData,
 		@Res() res: Response
 	) {
@@ -77,7 +60,10 @@ export class CategoryController {
 			return
 		}
 
-		const result = await this.categoryService.reorder(body, session.auth.userId)
+		const result = await this.categoryGroupService.reorder(
+			body,
+			session.auth.userId
+		)
 
 		if (result === null) {
 			return res
@@ -89,18 +75,18 @@ export class CategoryController {
 	}
 
 	@Put(':id')
-	async put(
-		@Param('id') id: string,
-		@Body() body: UpdateCategoryDto,
+	async edit(
 		@Session() session: SessionData,
-		@Res() res: Response
+		@Res() res: Response,
+		@Param('id') id: string,
+		@Body() body: UpdateCategoryGroupDto
 	) {
 		if (session.auth === undefined || !session.auth.authorized) {
 			res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' })
 			return
 		}
 
-		const result = await this.categoryService.edit(
+		const result = await this.categoryGroupService.edit(
 			id,
 			body,
 			session.auth.userId
@@ -113,5 +99,21 @@ export class CategoryController {
 		}
 
 		res.status(HttpStatus.OK).json(result)
+	}
+
+	@Delete(':id')
+	async delete(
+		@Session() session: SessionData,
+		@Res() res: Response,
+		@Param('id') id: string
+	) {
+		if (session.auth === undefined || !session.auth.authorized) {
+			res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' })
+			return
+		}
+
+		res
+			.status(HttpStatus.OK)
+			.json(await this.categoryGroupService.delete(id, session.auth.userId))
 	}
 }