+++ src/category-group/category-group.interface.ts
@@ -1,0 +1,5 @@
+import { HydratedDocument } from 'mongoose'
+import { Group } from './category-group.schema'
+
+// TODO: Renamed to ICategoryGroup
+export type IGroup = HydratedDocument<Group>