+++ src/category-group/dto/create-category-group.dto.ts
@@ -1,0 +1,19 @@
+import { ApiProperty } from '@nestjs/swagger'
+import { IsNotEmpty, IsNumber, IsString } from 'class-validator'
+
+export class CreateCategoryGroupDto {
+	@ApiProperty()
+	@IsNotEmpty()
+	@IsString()
+	name: string
+
+	@ApiProperty()
+	@IsNotEmpty()
+	@IsString()
+	color: string
+
+	@ApiProperty()
+	@IsNotEmpty()
+	@IsNumber()
+	order: number
+}